feat: add recommendation modal and all recommend button components, enhance state management for recommendations
