+++ src/domains/situation-create/store/use-recommend-form-store.ts
@@ -1,7 +1,30 @@
-import { TRecommendationCategories } from "../reducer/types";
+import { create } from "zustand";
+import { devtools } from "zustand/middleware";
 
 type RecommendFormState = {
+  isModalOpen: boolean;
   hasRequestedAllRecommendations: boolean;
-  isAllRecLoading: boolean;
-  currentRecommendLoading: keyof TRecommendationCategories | "all" | null;
+  setIsModalOpen: (flag: boolean) => void;
 };
+
+const initialState = {
+  isModalOpen: false,
+  hasRequestedAllRecommendations: false,
+};
+
+const useRecommendFormStore = create<RecommendFormState>()(
+  devtools(
+    (set) => ({
+      ...initialState,
+
+      setIsModalOpen: (flag) => {
+        set({
+          isModalOpen: flag,
+        });
+      },
+    }),
+    { name: "Recommend-Form" }
+  )
+);
+
+export default useRecommendFormStore;
